+++ inst/LDS/LDS_C12.R
@@ -1,6 +1,0 @@
-
-
-# Page 666, Question (Large Data Set 'SMOKING')
-head(SMOKING_raw <- read.csv('data/LDS_C12_SMOKING.csv'))
-
-# dont know what are the columns..  do not use this example