+++ inst/example/Chapter11.R
@@ -1,4 +1,12 @@
 library(DanielBiostatistics10th)
+# To run a line of code, use shortcut
+# Command + Enter: Mac and RStudio Cloud
+# Control + Enter: Windows, Mac and RStudio Cloud
+# To clear the console
+# Control + L: Mac and RStudio Cloud
+
+
+
 
 # Page 540, Example 11.1.1
 # data not provided by the publisher