+++ inst/example/Chapter9.R
@@ -1,4 +1,11 @@
 library(DanielBiostatistics10th)
+# To run a line of code, use shortcut
+# Command + Enter: Mac and RStudio Cloud
+# Control + Enter: Windows, Mac and RStudio Cloud
+# To clear the console
+# Control + L: Mac and RStudio Cloud
+
+
 
 # Page 417, Example 9.3.1
 # .. Correlation