+++ inst/example/LDS_C12.R
@@ -1,0 +1,14 @@
+library(DanielBiostatistics10th)
+# To run a line of code, use shortcut
+# Command + Enter: Mac and RStudio Cloud
+# Control + Enter: Windows, Mac and RStudio Cloud
+# To clear the console
+# Control + L: Mac and RStudio Cloud
+
+
+
+
+# Page 666, Question (Large Data Set 'SMOKING')
+head(SMOKING_raw <- read.csv('data/LDS_C12_SMOKING.csv'))
+
+# dont know what are the columns..  do not use this example