+++ inst/example/LDS_chp6.R
@@ -1,3 +1,10 @@
+library(DanielBiostatistics10th)
+# To run a line of code, use shortcut
+# Command + Enter: Mac and RStudio Cloud
+# Control + Enter: Windows, Mac and RStudio Cloud
+# To clear the console
+# Control + L: Mac and RStudio Cloud
+
 
 
 # Large Data Sets in Chapter 6, Page 210