+++ inst/example/LDS_chp7.R
@@ -1,3 +1,12 @@
+library(DanielBiostatistics10th)
+# To run a line of code, use shortcut
+# Command + Enter: Mac and RStudio Cloud
+# Control + Enter: Windows, Mac and RStudio Cloud
+# To clear the console
+# Control + L: Mac and RStudio Cloud
+
+
+
 
 # Large Data Sets in Chapter 7, Page 299