+++ NEWS.md
@@ -1,5 +1,9 @@
+# DanielBiostatistics10th 0.1.2
+Rename parameter `phat` of ?prop_test_CLT to `xbar`
+Add function ?print_freqs ?print_OE
+
 # DanielBiostatistics10th 0.1.1
-Remove package raster import.
+Remove package \pkg{raster} import.
 
 # DanielBiostatistics10th 0.1.0
 First release.